refactor: change name file utils to translate-utils
